+++ kernel/src/net/real_socket.h
@@ -49,13 +49,13 @@ typedef struct socket_handle {
 
 typedef struct real_socket_socket {
     int domain;
-    int (*socket)(int domain, int type, int protocol);
+    uint64_t (*socket)(int domain, int type, int protocol);
 } real_socket_socket_t;
 
 #define MAX_SOCKETS_NUM 16
 
 void regist_socket(int domain,
-                   int (*socket)(int domain, int type, int protocol));
+                   uint64_t (*socket)(int domain, int type, int protocol));
 
 extern real_socket_socket_t *real_sockets[MAX_SOCKETS_NUM];
 extern int socket_num;
